implemented the module as a function
