+++ docker-compose.yaml
@@ -151,14 +151,6 @@ services:
     restart: always
     volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
 
-  karton-domain_scanner:
-    <<: *artemis-build-or-image
-    command: "python3 -m artemis.modules.domain_scanner"
-    depends_on: [karton-system, karton-logger]
-    env_file: .env
-    restart: always
-    volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
-
   karton-drupal_scanner:
     <<: *artemis-build-or-image
     command: "python3 -m artemis.modules.drupal_scanner"
